code fix

+++ current/app/src/drvInficon.h
@@ -163,7 +163,7 @@ public:
 	asynStatus parseInt32(const char *jsonData, epicsInt32 *value, commandType_t commandType);
     asynStatus parseUInt32(const char *jsonData, epicsUInt32 *value, commandType_t commandType);
     asynStatus parseFloat64(const char *jsonData, epicsFloat64 *value, commandType_t commandType);
-    asynStatus parseString(const char *jsonData, char *data, int *dataLen, commandType_t commandType);
+    asynStatus parseString(const char *jsonData, char *data, size_t *dataLen, commandType_t commandType);
     asynStatus parseScan(const char *jsonData, double *scanValues, int *scanSize, int *scannum);
 	asynStatus verifyConnection();   // Verify connection using asynUser //Return asynSuccess for connect
     bool inficonExiting_;
